Add Bottom of the Well and Ganon's Castle MQ checks
The BotW checks look strange since there are actually 12
Wonder Items in the dungeon, but several of them have identical
positions and don't set any flags, meaning the code as it is
is incapable of distinguishing them. Several even have identical
params on top of that, making it even harder to detect which
is which.